+++ src/ng-app/app/cpu-dashboard/cpu-dashboard.component.ts
@@ -1,5 +1,5 @@
 /*!
- * Copyright (c) 2019 TUXEDO Computers GmbH <tux@tuxedocomputers.com>
+ * Copyright (c) 2019-2020 TUXEDO Computers GmbH <tux@tuxedocomputers.com>
  *
  * This file is part of TUXEDO Control Center.
  *
@@ -16,11 +16,17 @@
  * You should have received a copy of the GNU General Public License
  * along with TUXEDO Control Center.  If not, see <https://www.gnu.org/licenses/>.
  */
-import { Component, OnInit, OnDestroy } from '@angular/core';
+import { Component, OnInit, OnDestroy, ViewEncapsulation } from '@angular/core';
 import { ILogicalCoreInfo, IGeneralCPUInfo, SysFsService, IPstateInfo } from '../sys-fs.service';
 import { Subscription } from 'rxjs';
 import { UtilsService } from '../utils.service';
 import { TccDBusClientService, IDBusFanData } from '../tcc-dbus-client.service';
+import { ITccProfile } from 'src/common/models/TccProfile';
+import { StateService } from '../state.service';
+import { Router } from '@angular/router';
+import { ConfigService } from '../config.service';
+
+import { I18n } from '@ngx-translate/i18n-polyfill';
 
 @Component({
   selector: 'app-cpu-dashboard',
@@ -40,21 +46,71 @@ export class CpuDashboardComponent implements OnInit, OnDestroy {
   public activeScalingGovernors: string[];
   public activeEnergyPerformancePreference: string[];
 
+  public avgCpuFreq: number;
+  public avgCpuFreqData;
+
   public fanData: IDBusFanData;
+  public gaugeGPUTemp: number;
+  public gaugeGPUSpeed: number;
+  public hasGPUTemp = false;
+
+  public activeProfile: ITccProfile;
+  public isCustomProfile: boolean;
+
+  public fanProfileProgressMap: Map<string, number>;
 
   private subscriptions: Subscription = new Subscription();
 
   constructor(
     private sysfs: SysFsService,
     private utils: UtilsService,
-    private tccdbus: TccDBusClientService
+    private tccdbus: TccDBusClientService,
+    private state: StateService,
+    private router: Router,
+    private config: ConfigService,
+    private i18n: I18n
   ) { }
 
   ngOnInit() {
     this.subscriptions.add(this.sysfs.generalCpuInfo.subscribe(cpuInfo => { this.cpuInfo = cpuInfo; }));
     this.subscriptions.add(this.sysfs.logicalCoreInfo.subscribe(coreInfo => { this.cpuCoreInfo = coreInfo; this.updateFrequencyData(); }));
     this.subscriptions.add(this.sysfs.pstateInfo.subscribe(pstateInfo => { this.pstateInfo = pstateInfo; }));
-    this.subscriptions.add(this.tccdbus.fanData.subscribe(fanData => { this.fanData = fanData; }));
+    this.subscriptions.add(this.tccdbus.fanData.subscribe(fanData => {
+      this.fanData = fanData;
+      let avgTemp: number;
+      let avgSpeed: number;
+      const temp1 = this.fanData.gpu1.temp.data.value;
+      const temp2 = this.fanData.gpu2.temp.data.value;
+      const speed1 = this.fanData.gpu1.speed.data.value;
+      const speed2 = this.fanData.gpu2.speed.data.value;
+      if (temp1 !== 1 && temp2 !== 1) {
+        avgTemp = (temp1 + temp2) / 2;
+        avgSpeed = (speed1 + speed2) / 2;
+      } else if (temp1 !== 1) {
+        avgTemp = temp1;
+        avgSpeed = speed1;
+      } else {
+        // This covers two cases, temp2 having a valid temperature, and not having GPU temperature at all
+        avgTemp = temp2;
+        avgSpeed = speed2;
+      }
+      this.hasGPUTemp = avgTemp !== 1;
+      this.gaugeGPUTemp = Math.round(avgTemp);
+      this.gaugeGPUSpeed = Math.round(avgSpeed);
+    }));
+    this.subscriptions.add(this.state.activeProfile.subscribe(profile => {
+      this.activeProfile = profile;
+      this.isCustomProfile = this.config.getCustomProfileByName(this.activeProfile.name) !== undefined;
+    }));
+
+    const fanProfileNames = this.config.getFanProfiles().map(fanProfile => fanProfile.name);
+    this.fanProfileProgressMap = new Map();
+    const fanProfileProgressStep = 100.0 / fanProfileNames.length;
+    let progressValue = 0;
+    for (const fanProfileName of fanProfileNames) {
+      progressValue += fanProfileProgressStep;
+      this.fanProfileProgressMap.set(fanProfileName, progressValue);
+    }
   }
 
   ngOnDestroy(): void {
@@ -86,9 +142,35 @@ export class CpuDashboardComponent implements OnInit, OnDestroy {
         this.activeScalingDrivers.push(core.scalingDriver);
       }
     }
+
+    // Calculate average frequency over the logical cores
+    const freqSum: number =
+      this.cpuCoreInfo
+      .map(info => info.scalingCurFreq)
+      .reduce((sum, currentFreq) => sum + currentFreq, 0);
+    this.avgCpuFreq = freqSum / this.cpuCoreInfo.length;
+    this.avgCpuFreqData = [{ name: 'CPU frequency', value: this.avgCpuFreq }];
   }
 
-  public formatFrequency(frequency: number): string {
+  public formatFrequency = (frequency: number): string => {
     return this.utils.formatFrequency(frequency);
+  }
+
+  public gaugeFreqFormat: (value: number) => string = (value) => {
+    return this.utils.formatFrequency(value);
+  }
+
+  public gaugeOnOffFormat: (value: number) => string = (value) => {
+    if (value === 0) {
+      return this.i18n({ value: 'off'});
+    } else {
+      return this.i18n({ value: 'on'});
+    }
+  }
+
+  public goToProfileEdit(profile: ITccProfile): void {
+    if (profile !== undefined) {
+      this.router.navigate(['profile-manager', profile.name]);
+    }
   }
 }