+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Changelog
 
-## [0.7.3] - 2019-01-02
+## [0.8.0] - 2020-03-06
+## Added / Changed
+- New control through tuxedo-wmi module
+- Most of the new design implemented
+
+## [0.7.4] - 2019-12-09
+## Added
+- Theme switching
+- Light/dark theme for testing & toggling
+
+## [0.7.3] - 2019-12-02
 ## Changed
 - Preset profile changes
 - Removed *Fast and furious*